refactor: split main.dart's _AppShellState into 3 themed extensions
main.dart was 1255 LoC. The bulk lived in _AppShellState. Extract three
contiguous themed blocks into part files under lib/app_shell/, each an
extension on _AppShellState:

- share_intent.dart: Android share intent handling (~105 LoC)
- drive_db_ops.dart: Drive backup/restore + DB import/wipe + import/
  export dialog (~254 LoC)
- settings_dialog.dart: the settings dialog (~219 LoC)

Two small adjustments to keep the extensions valid:
- `_localeOptions` static const moved to library-top-level — Dart
  extensions can't reach unqualified static members of the extended type.
- dispose() and one setState() stay in _AppShellState's class body —
  they can't go in extensions (override / @protected on State).

main.dart: 1255 → 687 LoC (−45%). All four suites green.
